Script is ready
